# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "欺骗防御接口文档",
        "contact": {
            "name": "akita.tian",
            "email": "chronoyq@163.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8082",
    "paths": {
...
        "/api/public/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"token\":\"\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"code\":2001,\"msg\":\"账号或是密码错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/public/signup": {
            "post": {
                "description": "注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"token\":\"\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/password": {
            "put": {
                "description": "修改密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "NewPassword",
                        "name": "NewPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_handler.PasswordChangePayload"
                        }
                    },
                    {
                        "description": "RepeatPassword",
                        "name": "RepeatPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_handler.PasswordChangePayload"
                        }
                    },
                    {
                        "description": "OldPassword",
                        "name": "OldPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_handler.PasswordChangePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"内部异常\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/webhook": {
            "get": {
                "description": "获取钉钉webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取钉钉webhook",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "设置钉钉webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "设置钉钉webhook",
                "parameters": [
                    {
                        "description": "WebHook",
                        "name": "WebHook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webhook_handler.WebHookConfigPayload"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"请求参数错误\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
...
        "user_handler.PasswordChangePayload": {
            "type": "object",
            "required": [
                "NewPassword",
                "OldPassword",
                "RepeatPassword"
            ],
            "properties": {
                "NewPassword": {
                    "type": "string"
                },
                "OldPassword": {
                    "type": "string"
                },
                "RepeatPassword": {
                    "type": "string"
                }
            }
        },
        "webhook_handler.WebHookConfigPayload": {
            "type": "object",
            "required": [
                "WebHook"
            ],
            "properties": {
                "WebHook": {
                    "description": "钉钉接口",
                    "type": "string"
                }
            }
        }
    }
}